Чистка консоли рендера от ненужных команд

+++ xrRender/r_rendertarget.cpp
@@ -324,7 +324,7 @@ CRenderTarget::CRenderTarget()
 		break;
 	}
 
-	u32 w = dwWidth * BloomResolutionMultiplier, h = dwHeight * BloomResolutionMultiplier;
+	u32 w = u32(dwWidth * BloomResolutionMultiplier), h = u32(dwHeight * BloomResolutionMultiplier);
 	u32 fvf_build = D3DFVF_XYZRHW | D3DFVF_TEX4 | D3DFVF_TEXCOORDSIZE2(0) | D3DFVF_TEXCOORDSIZE2(1) |
 					D3DFVF_TEXCOORDSIZE2(2) | D3DFVF_TEXCOORDSIZE2(3);
 	u32 fvf_filter = (u32)D3DFVF_XYZRHW | D3DFVF_TEX8 | D3DFVF_TEXCOORDSIZE4(0) | D3DFVF_TEXCOORDSIZE4(1) |
